requirements.txt
fastapi
uvicorn
pandas
numpy
sentence-transformers
faiss-cpu
groq
scikit-learn
python-dotenv
pydantic-settings
jinja2
aiofiles
bcrypt